+++ gg-pingpong-api/src/main/java/gg/pingpong/api/user/user/controller/UserController.java
@@ -31,8 +31,7 @@ import gg.auth.argumentresolver.Login;
 import gg.data.pingpong.game.type.Mode;
 import gg.data.user.type.OauthType;
 import gg.data.user.type.RoleType;
-import gg.pingpong.api.global.security.cookie.CookieUtil;
-import gg.pingpong.api.global.security.jwt.utils.TokenHeaders;
+import gg.pingpong.api.global.jwt.utils.TokenHeaders;
 import gg.pingpong.api.user.user.controller.request.UserModifyRequestDto;
 import gg.pingpong.api.user.user.controller.request.UserProfileImageRequestDto;
 import gg.pingpong.api.user.user.controller.response.UserAttendanceResponseDto;
@@ -53,6 +52,7 @@ import gg.pingpong.api.user.user.dto.UserTextColorDto;
 import gg.pingpong.api.user.user.service.UserAuthenticationService;
 import gg.pingpong.api.user.user.service.UserCoinService;
 import gg.pingpong.api.user.user.service.UserService;
+import gg.utils.cookie.CookieUtil;
 import gg.utils.dto.PageRequestDto;
 import gg.utils.exception.user.KakaoOauth2AlreadyExistException;
 import io.swagger.v3.oas.annotations.Parameter;
