+++ gg-pingpong-api/src/main/java/gg/pingpong/api/user/user/service/UserAuthenticationService.java
@@ -3,7 +3,7 @@ package gg.pingpong.api.user.user.service;
 import org.springframework.stereotype.Service;
 
 import gg.auth.utils.AuthTokenProvider;
-import gg.pingpong.api.global.security.jwt.repository.JwtRedisRepository;
+import gg.repo.user.JwtRedisRepository;
 import gg.utils.exception.user.TokenNotValidException;
 import lombok.RequiredArgsConstructor;
