+++ gg-pingpong-api/src/test/java/gg/pingpong/api/user/rank/controller/RankV2ControllerTest.java
@@ -23,8 +23,8 @@ import gg.auth.UserDto;
 import gg.auth.utils.AuthTokenProvider;
 import gg.data.user.User;
 import gg.pingpong.api.global.config.WebConfig;
+import gg.pingpong.api.global.jwt.utils.TokenAuthenticationFilter;
 import gg.pingpong.api.global.security.config.SecurityConfig;
-import gg.pingpong.api.global.security.jwt.utils.TokenAuthenticationFilter;
 import gg.pingpong.api.global.utils.querytracker.LoggingInterceptor;
 import gg.pingpong.api.user.rank.controller.response.ExpRankPageResponseDto;
 import gg.pingpong.api.user.rank.controller.response.RankPageResponseDto;
