+++ gg-pingpong-api/src/test/java/gg/pingpong/api/user/tournament/controller/TournamentControllerMvcTest.java
@@ -23,8 +23,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import gg.auth.UserDto;
 import gg.auth.config.AuthWebConfig;
 import gg.pingpong.api.global.config.WebConfig;
+import gg.pingpong.api.global.jwt.utils.TokenAuthenticationFilter;
 import gg.pingpong.api.global.security.config.SecurityConfig;
-import gg.pingpong.api.global.security.jwt.utils.TokenAuthenticationFilter;
 import gg.pingpong.api.global.utils.querytracker.LoggingInterceptor;
 import gg.pingpong.api.user.tournament.controller.request.TournamentFilterRequestDto;
 import gg.pingpong.api.user.tournament.controller.response.TournamentGameListResponseDto;
