+++ gg-repo/src/main/java/gg/repo/user/JwtRedisRepository.java
@@ -1,4 +1,4 @@
-package gg.pingpong.api.global.security.jwt.repository;
+package gg.repo.user;
 
 import java.util.concurrent.TimeUnit;
